Additional critera for prrojections equality?
If central meridian, spheroid member values and unit member values are
equal, is that a sufficient condition (but not necessary) for two
projections to be considered equivalent? If so, I propose to change the
Projections.Equals method.

Also, if the authority codes are the same for two projections, it should
be safe to conclude that they are equal.