Fix Proposal for Symbology FilterExpression
As described here: #1160 there is an issue when using a literal asterisk (ex "BOGUS*" or "**LENTL") for the Symbology FilterExpression. An exception occurs when using the FilterExpression through code or in the DemoMap when importing a DSPX file. Oddly enough the exception doesnt occur when setting the FilterExpression through the Symbology editor.