Screen referenced labels
Since the DotSpatial framework currently does not provide a mechanism for screen referenced labels the best we can do in the meantime is to allow the developer to pass in Label objects that are controls placed on top of the map. This code overloads the Map.Snapshot routine to allow .NET labels to be included in the exported map bitmap.