Symbology FilterExpression with Literal Wildcard
Fix Proposal for Symbology FilterExpression

As described here: #1160  https://github.com/DotSpatial/DotSpatial/issues/1160 there is an issue when using a literal asterisk (ex "BOGUS*" or "**LENTL") for the Symbology FilterExpression. An exception occurs when using the FilterExpression through code or in the DemoMap when importing a DSPX file. Oddly enough the exception doesnt occur when setting the FilterExpression through the Symbology editor. Why would this be checked on load but not when changing symbology at a later time?

https://docs.microsoft.com/en-us/dotnet/api/system.data.datacolumn.expression?redirectedfrom=MSDN&view=netframework-4.7.2#System_Data_DataColumn_Expression

"Both the * and % can be used interchangeably for wildcard characters in a LIKE comparison. If the string in a LIKE clause contains a * or %, those characters should be enclosed in brackets ([])."

Without this fix the framework will parse out the lietral asterisk as a wildcard and the result is an empty string which will cause an unhandled exception.