Drawing order of labels is given precedence from top layer to bottom layer
Layers higher in the draw hierarchy should be drawn on top, and so should their labels.
In other words the drawing order of labels is given precedence from top layer to bottom layer.

Not my logic but wanted to add it to my pull requests for practice. I have other pull requests so this is just something I was interested in.

Based on issue #1226 Drawing order of labels incorrect