Labeler Crash with Mercator or Orthographic projection
When using a map projection of Mercator or Orthographic (possibly others as well) and labeling features there is a small bug that I have been trying to squash all weekend. I believe it is because when the reprojection of the polygons occurs there are some inherent values of NaN for x or y or both. When the labeler tries to determine the centroid of the said features it results in an exception.