Multiple reprojections lose original vertices
Reprojects all layers in the map frame so that they use new projection. Before they are reprojected the layers will get reloaded from the original source if available. This is useful when doing multiple reprojections on a layer as the reprojection will occur on the original vertices and not the reprojected vertices. For certain workflows multiple reprojections will tend to lose the original vertices and result in unexpected coordinates. This method attempts to alleviate the issue. This method also includes drawing layers, although they may lose their original vertices since they don't have a source from where we can reload.