Polygon label on perimeter of feature
This modifciation introduces a label placement around the perimeter of a polygon feature. When the user emplys this label placement method the code basically treats the outer ring of a polygon feature as a series of line segments and uses most of the polyline logic.

Implemented was a new variation of the rotation angle. The current code calculates it from the segment and uses ATAN. That works for all the calculations except when it comes to the inside/outside of a polygon since it loses the directionality. To account for that a new calcuation was introduced with ATAN2 and called theta.

The polygon requires this angle with directionality so that we can determine which side is inside or outside the polygon. The only place theta is used is in the Position method and it doesn't get used for text rotation. All we need theta for is to determine which direction to offset the text relative to the segment. Basically logic to determine if the text is above or below the line.

The DotSpatial.Controls.MapLabelLayer is great code but could use a refactoring at some point. Of note is the multiple calls to the GetSegment method which could be needless overhead. I could envision that DotSpatil could be modified to allow different labeling engines much like a plugin.

For now I think this new label placement method is a good starting point for further work like optimal placement on the perimeter rather than along the first, midlle, last, or longest segment. This can be seen in the sample project where the label is always placed on the longest segment. This is only in the UI as the developer can set them as they wish in code. This is due to LabelPlacementMethod and LineLabelPlacementMethod not being mutually exclusive. Perhaps a new SegmentLabelPlacementMethod would resolve this.