Auto display children of MapGroup is now an option of the MapFrame (#1254)
Auto display children of MapGroup is now an option of the MapFrame. What this means is we can now set a value indicating whether or not the visibility of the children of a group are triggered by the checked status of the group. If this is false (the default) then this will force all the layers in this group to become visible. if this set to true then the visbility of each layer is dependent on the status of every one of its parent group. I looked for the best place to put this property and settled on the MapFrame.